requirements.txt
streamlit
google-generativeai
tkinter